+++ src/utils/msgIdFn.ts
@@ -1,5 +1,5 @@
 import { sha256 } from 'multiformats/hashes/sha2'
-import type { Message } from '@libp2p/interface-pubsub'
+import type { Message } from '@libp2p/interface/pubsub'
 import { msgId } from '@libp2p/pubsub/utils'
 
 /**