+++ test/heartbeat.spec.ts
@@ -1,9 +1,9 @@
 import { expect } from 'aegir/chai'
 import { GossipsubHeartbeatInterval } from '../src/constants.js'
 import { pEvent } from 'p-event'
-import { createComponents, GossipSubAndComponents } from './utils/create-pubsub.js'
-import { stop } from '@libp2p/interfaces/startable'
-import { mockNetwork } from '@libp2p/interface-mocks'
+import { createComponents, type GossipSubAndComponents } from './utils/create-pubsub.js'
+import { stop } from '@libp2p/interface/startable'
+import { mockNetwork } from '@libp2p/interface-compliance-tests/mocks'
 
 describe('heartbeat', () => {
   let node: GossipSubAndComponents