+++ test/mesh.spec.ts
@@ -2,9 +2,9 @@ import { expect } from 'aegir/chai'
 import delay from 'delay'
 import { GossipsubDhi } from '../src/constants.js'
 import type { GossipSub } from '../src/index.js'
-import { connectAllPubSubNodes, createComponentsArray, GossipSubAndComponents } from './utils/create-pubsub.js'
-import { stop } from '@libp2p/interfaces/startable'
-import { mockNetwork } from '@libp2p/interface-mocks'
+import { connectAllPubSubNodes, createComponentsArray, type GossipSubAndComponents } from './utils/create-pubsub.js'
+import { stop } from '@libp2p/interface/startable'
+import { mockNetwork } from '@libp2p/interface-compliance-tests/mocks'
 import { pEvent } from 'p-event'
 
 describe('mesh overlay', () => {